Add VolumeCroppingControlTool to tools index export
