+++ packages/tools/src/index.ts
@@ -29,6 +29,7 @@ import {
   PanTool,
   TrackballRotateTool,
   VolumeCroppingTool,
+  VolumeCroppingControlTool,
   DragProbeTool,
   WindowLevelTool,
   ZoomTool,
@@ -108,6 +109,7 @@ export {
   SegmentBidirectionalTool,
   TrackballRotateTool,
   VolumeCroppingTool,
+  VolumeCroppingControlTool,
   DragProbeTool,
   WindowLevelTool,
   WindowLevelRegionTool,
