+++ packages/tools/src/tools/VolumeCroppingTool.ts
@@ -21,22 +21,13 @@ import {
   triggerEvent,
   eventTarget,
 } from '@cornerstonejs/core';
+import { Enums as toolsEnums } from '@cornerstonejs/tools';
 
 import {
   getToolGroup,
   getToolGroupForViewport,
 } from '../store/ToolGroupManager';
 
-import {
-  addAnnotation,
-  getAnnotations,
-  removeAnnotation,
-} from '../stateManagement/annotation/annotationState';
-
-import {
-  drawCircle as drawCircleSvg,
-  drawLine as drawLineSvg,
-} from '../drawingSvg';
 import { state } from '../store/state';
 import { Events } from '../enums';
 import { getViewportIdsWithToolToRender } from '../utilities/viewportFilters';
@@ -44,7 +35,6 @@ import {
   resetElementCursor,
   hideElementCursor,
 } from '../cursors/elementCursor';
-import liangBarksyClip from '../utilities/math/vec2/liangBarksyClip';
 
 import * as lineSegment from '../utilities/math/line';
 import type {
@@ -55,7 +45,6 @@ import type {
   PublicToolProps,
   ToolProps,
   InteractionTypes,
-  SVGDrawingHelper,
 } from '../types';
 import { isAnnotationLocked } from '../stateManagement/annotation/annotationLocking';
 import triggerAnnotationRenderForViewportIds from '../utilities/triggerAnnotationRenderForViewportIds';
@@ -120,11 +109,9 @@ class VolumeCroppingTool extends AnnotationTool {
     defaultToolProps: ToolProps = {
       supportedInteractionTypes: ['Mouse'],
       configuration: {
-        shadow: true,
         // renders a colored circle on top right of the viewports whose color
         // matches the color of the reference line
         viewportIndicators: false,
-
         viewportIndicatorsConfig: {
           radius: 5,
           x: null,
@@ -154,6 +141,13 @@ class VolumeCroppingTool extends AnnotationTool {
           handleRadius: 9,
         },
         initialCropFactor: 0.2,
+        sphereColors: {
+          x: [0.0, 1.0, 0.0], // Green for X
+          y: [1.0, 1.0, 0.0], // Yellow for Y
+          z: [1.0, 0.0, 0.0], // Red for Z
+          default: [0.0, 0.0, 1.0], // Blue as fallback
+        },
+        sphereRadius: 10,
       },
     }
   ) {
@@ -171,82 +165,8 @@ class VolumeCroppingTool extends AnnotationTool {
     this.picker.initializePickList();
   }
 
-  /**
-   * Gets the camera from the viewport, and adds  annotation for the viewport
-   * to the annotationManager. If any annotation is found in the annotationManager, it
-   * overwrites it.
-   * @param viewportInfo - The viewportInfo for the viewport to add the crosshairs
-   * @returns viewPlaneNormal and center of viewport canvas in world space
-   */
-  initializeViewport = ({
-    renderingEngineId,
-    viewportId,
-  }: Types.IViewportId): {
-    normal: Types.Point3;
-    point: Types.Point3;
-  } => {
-    const enabledElement = getEnabledElementByIds(
-      viewportId,
-      renderingEngineId
-    );
-    if (!enabledElement) {
-      return;
-    }
-    const { FrameOfReferenceUID, viewport } = enabledElement;
-    const { element } = viewport;
-    const { position, focalPoint, viewPlaneNormal } = viewport.getCamera();
-
-    // Check if there is already annotation for this viewport
-    let annotations = this._getAnnotations(enabledElement);
-    annotations = this.filterInteractableAnnotationsForElement(
-      element,
-      annotations
-    );
-
-    if (annotations?.length) {
-      // If found, it will override it by removing the annotation and adding it later
-      removeAnnotation(annotations[0].annotationUID);
-    }
-
-    const annotation = {
-      highlighted: false,
-      metadata: {
-        cameraPosition: <Types.Point3>[...position],
-        cameraFocalPoint: <Types.Point3>[...focalPoint],
-        FrameOfReferenceUID,
-        toolName: this.getToolName(),
-      },
-      data: {
-        handles: {
-          //  rotationPoints: [], // rotation handles, used for rotation interactions
-          //  slabThicknessPoints: [], // slab thickness handles, used for setting the slab thickness
-          toolCenter: this.toolCenter,
-        },
-        activeOperation: null, // 0 translation, 1 rotation handles, 2 slab thickness handles
-        activeViewportIds: [], // a list of the viewport ids connected to the reference lines being translated
-        viewportId,
-        referenceLines: [], // set in renderAnnotation
-        clippingPlanes: [], // clipping planes for the viewport
-        clippingPlaneReferenceLines: [], // clipping planes for the viewport
-      },
-    };
-    if (viewport.type !== Enums.ViewportType.VOLUME_3D) {
-      addAnnotation(annotation, element);
-      return {
-        normal: viewPlaneNormal,
-        point: viewport.canvasToWorld([
-          viewport.canvas.clientWidth / 2,
-          viewport.canvas.clientHeight / 2,
-        ]),
-      };
-    } else {
-      return;
-    }
-  };
-
   _getViewportsInfo = () => {
     const viewports = getToolGroup(this.toolGroupId).viewportsInfo;
-
     return viewports;
   };
 
@@ -258,94 +178,29 @@ class VolumeCroppingTool extends AnnotationTool {
     // and update the reference points accordingly.
     this._unsubscribeToViewportNewVolumeSet(viewportsInfo);
     this._subscribeToViewportNewVolumeSet(viewportsInfo);
-
-    this._computeToolCenter(viewportsInfo);
+    this._initialize3DViewports(viewportsInfo);
   }
 
   onSetToolPassive() {
     const viewportsInfo = this._getViewportsInfo();
-
-    this._computeToolCenter(viewportsInfo);
+    this._initialize3DViewports(viewportsInfo);
   }
 
   onSetToolEnabled() {
     const viewportsInfo = this._getViewportsInfo();
-
-    //   this._computeToolCenter(viewportsInfo);
+    this._initialize3DViewports(viewportsInfo);
   }
 
   onSetToolDisabled() {
     const viewportsInfo = this._getViewportsInfo();
-
     this._unsubscribeToViewportNewVolumeSet(viewportsInfo);
-
-    // Crosshairs annotations in the state
-    // has no value when the tool is disabled
-    // since viewports can change (zoom, pan, scroll)
-    // between disabled and enabled/active states.
-    // so we just remove the annotations from the state
-    viewportsInfo.forEach(({ renderingEngineId, viewportId }) => {
-      const enabledElement = getEnabledElementByIds(
-        viewportId,
-        renderingEngineId
-      );
-
-      if (!enabledElement) {
-        return;
-      }
-
-      const annotations = this._getAnnotations(enabledElement);
-
-      if (annotations?.length) {
-        annotations.forEach((annotation) => {
-          removeAnnotation(annotation.annotationUID);
-        });
-      }
-    });
   }
-
-  resetCrosshairs = () => {
-    const viewportsInfo = this._getViewportsInfo();
-    for (const viewportInfo of viewportsInfo) {
-      const { viewportId, renderingEngineId } = viewportInfo;
-      const enabledElement = getEnabledElementByIds(
-        viewportId,
-        renderingEngineId
-      );
-      const viewport = enabledElement.viewport as Types.IVolumeViewport;
-      const resetPan = true;
-      const resetZoom = true;
-      const resetToCenter = true;
-      const resetRotation = true;
-      const suppressEvents = true;
-      viewport.resetCamera({
-        resetPan,
-        resetZoom,
-        resetToCenter,
-        resetRotation,
-        suppressEvents,
-      });
-      (viewport as Types.IVolumeViewport).resetSlabThickness();
-      const { element } = viewport;
-      let annotations = this._getAnnotations(enabledElement);
-      annotations = this.filterInteractableAnnotationsForElement(
-        element,
-        annotations
-      );
-      if (annotations.length) {
-        removeAnnotation(annotations[0].annotationUID);
-      }
-      viewport.render();
-    }
-
-    this._computeToolCenter(viewportsInfo);
-  };
 
   addSphere(viewport, point, axis) {
     // Generate a unique UID for each sphere based on its axis and position
     const uid = `${axis}_${point.map((v) => Math.round(v)).join('_')}`;
     const sphereState = this.sphereStates.find((s) => s.uid === uid);
-    //   if (!sphereState) {
+    //  if (!sphereState) {
     const sphereSource = vtkSphereSource.newInstance();
     sphereSource.setCenter(point);
     sphereSource.setRadius(15);
@@ -368,15 +223,14 @@ class VolumeCroppingTool extends AnnotationTool {
       this.sphereStates[idx].point = point.slice();
       this.sphereStates[idx].sphereSource = sphereSource;
     }
-    /*
-      // Remove existing actor with this UID if present
-      const existingActors = viewport.getActors();
-      const existing = existingActors.find((a) => a.uid === uid);
-      if (existing) {
-        //return;
-        // viewport.removeActor(uid);
-      }
-*/
+
+    // Remove existing actor with this UID if present
+    const existingActors = viewport.getActors();
+    const existing = existingActors.find((a) => a.uid === uid);
+    if (existing) {
+      //viewport.removeActor(uid);
+      return;
+    }
 
     let color = [0.0, 1.0, 0.0];
     if (axis === 'z') {
@@ -384,22 +238,20 @@ class VolumeCroppingTool extends AnnotationTool {
     } else if (axis === 'x') {
       color = [1.0, 1.0, 0.0];
     }
-
     //const color = [0.0, 0.0, 1.0];
-
     sphereActor.getProperty().setColor(color);
-    //sphereActor.getProperty().setColor(0.0, 0.0, 1.0);
+
     /*
     const sphereColors = this.configuration.sphereColors || {};
-    const color = sphereColors[sphereState.axis] ||
+    const color = sphereColors[this.sphereStates[idx].axis] ||
       sphereColors.default || [0.0, 0.0, 1.0];
-    sphereActor.getProperty().setColor(...color);
+    sphereActor.getProperty().setColor(color);
 */
     sphereActor.setPickable(true);
     viewport.addActor({ actor: sphereActor, uid: uid });
     console.debug('added sphere: ', uid, viewport.getActors());
     viewport.render();
-    /*  } else {
+    /*   } else {
       // Only update the position and source, do not create new actor/source
       sphereState.point = point.slice();
       if (sphereState.sphereSource) {
@@ -411,64 +263,17 @@ class VolumeCroppingTool extends AnnotationTool {
       */
   }
 
-  computeToolCenter = () => {
-    const viewportsInfo = this._getViewportsInfo();
-    this._computeToolCenter(viewportsInfo);
-  };
+  _initialize3DViewports = (viewportsInfo): void => {
+    const [viewport3D] = viewportsInfo;
 
-  /**
-   * When activated, it initializes the crosshairs. It begins by computing
-   * the intersection of viewports associated with the crosshairs instance.
-   * When all three views are accessible, the intersection (e.g., crosshairs tool centre)
-   * will be an exact point in space; however, with two viewports, because the
-   * intersection of two planes is a line, it assumes the last view is between the centre
-   * of the two rendering viewports.
-   * @param viewportsInfo Array of viewportInputs which each item containing `{viewportId, renderingEngineId}`
-   */
-  _computeToolCenter = (viewportsInfo): void => {
-    // Todo: handle two same view viewport, or more than 3 viewports
-    const [firstViewport, secondViewport, thirdViewport, viewport3D] =
-      viewportsInfo;
-
-    // Initialize first viewport
-    const { normal: normal1, point: point1 } =
-      this.initializeViewport(firstViewport);
-
-    // Initialize second viewport
-    const { normal: normal2, point: point2 } =
-      this.initializeViewport(secondViewport);
-
-    let normal3 = <Types.Point3>[0, 0, 0];
-    let point3 = vec3.create();
-
-    // If there are three viewports
-    if (thirdViewport) {
-      ({ normal: normal3, point: point3 } =
-        this.initializeViewport(thirdViewport));
-    } else {
-      // If there are only two views (viewport) associated with the crosshairs:
-      // In this situation, we don't have a third information to find the
-      // exact intersection, and we "assume" the third view is looking at
-      // a location in between the first and second view centers
-      vec3.add(point3, point1, point2);
-      vec3.scale(point3, point3, 0.5);
-      vec3.cross(normal3, normal1, normal2);
-    }
-
-    // Planes of each viewport
-    const firstPlane = csUtils.planar.planeEquation(normal1, point1);
-    const secondPlane = csUtils.planar.planeEquation(normal2, point2);
-    const thirdPlane = csUtils.planar.planeEquation(normal3, point3);
-
-    this.initializeViewport(viewport3D);
-    // add clipping planes
     const renderingEngine = getRenderingEngine(viewport3D.renderingEngineId);
     const viewport = renderingEngine.getViewport(viewport3D.viewportId);
 
     const volumeActors = viewport.getActors();
     const imageData = volumeActors[0].actor.getMapper().getInputData();
     const dimensions = imageData.getDimensions();
-    console.debug('dimensions', dimensions);
+    const origin = imageData.getOrigin();
+    //console.debug('dimensions', dimensions);
     const spacing = imageData.getSpacing(); // [xSpacing, ySpacing, zSpacing]
     const worldDimensions = [
       Math.round(dimensions[0] * spacing[0]),
@@ -476,13 +281,24 @@ class VolumeCroppingTool extends AnnotationTool {
       Math.round(dimensions[2] * spacing[2]),
     ];
     //const worldDimensions = dimensions;
-    console.debug('worldDimensions', worldDimensions);
+    console.debug('worldDimensions', worldDimensions, origin);
+    //   /*
     const xMin = worldDimensions[0] * -0.5;
     const xMax = worldDimensions[0] * 0.5;
     const yMin = worldDimensions[1] * -0.5;
     const yMax = worldDimensions[1] * 0.5;
     const zMin = worldDimensions[2] * -0.5;
     const zMax = worldDimensions[2];
+    // */
+
+    /*
+    const xMin = origin[0] - worldDimensions[0] / 2;
+    const xMax = origin[0] + worldDimensions[0] / 2;
+    const yMin = origin[1] - worldDimensions[1] / 2;
+    const yMax = origin[1] + worldDimensions[1] / 2;
+    const zMin = origin[2] - (spacing[2] * (dimensions[2] - 1)) / 2;
+    const zMax = origin[2] + (spacing[2] * (dimensions[2] - 1)) / 2;
+*/
     const planes: vtkPlane[] = [];
     const cropFactor = 0.2;
     // X min plane (cuts everything left of xMin)
@@ -502,12 +318,16 @@ class VolumeCroppingTool extends AnnotationTool {
       origin: [0, yMax - worldDimensions[1] * cropFactor, 0],
       normal: [0, -1, 0],
     });
+
+    //     const sphereZminPoint = [(xMax + xMin) / 2, (yMax + yMin) / 2, -zMax];
+    //  const sphereZmaxPoint = [(xMax + xMin) / 2, (yMax + yMin) / 2, zMax / 4];
+
     const planeZmin = vtkPlane.newInstance({
       origin: [0, 0, -zMax],
       normal: [0, 0, 1],
     });
     const planeZmax = vtkPlane.newInstance({
-      origin: [0, 0, zMax],
+      origin: [0, 0, zMax / 4],
       normal: [0, 0, -1],
     });
 
@@ -532,39 +352,48 @@ class VolumeCroppingTool extends AnnotationTool {
 
     viewport.setOriginalClippingPlanes(originalPlanes);
 
+    const sphereXminPoint = [
+      xMin + worldDimensions[0] * cropFactor,
+      (yMax + yMin) / 2,
+      -220,
+    ];
+    const sphereXmaxPoint = [
+      xMax - worldDimensions[0] * cropFactor,
+      (yMax + yMin) / 2,
+      -220,
+    ];
+    const sphereYminPoint = [
+      (xMax + xMin) / 2,
+      yMin + worldDimensions[1] * cropFactor,
+      -220,
+    ];
+    const sphereYmaxPoint = [
+      (xMax + xMin) / 2,
+      yMax - worldDimensions[1] * cropFactor,
+      -220,
+    ];
+    /*
+    const sphereZminPoint = [
+      (xMax + xMin) / 2,
+      (yMax + yMin) / 2,
+      -300, //zMin + worldDimensions[2] * cropFactor,
+    ];
+    const sphereZmaxPoint = [
+      (xMax + xMin) / 2,
+      (yMax + yMin) / 2,
+      100, //zMax - worldDimensions[2] * cropFactor,
+    ];
+*/
+    const sphereZminPoint = [(xMax + xMin) / 2, (yMax + yMin) / 2, -zMax];
+    const sphereZmaxPoint = [(xMax + xMin) / 2, (yMax + yMin) / 2, zMax / 4];
+
     // this.addSphere(viewport, [xMin + worldDimensions[0] * cropFactor, 0, 0]);
-    this.addSphere(
-      viewport,
-      [xMin + worldDimensions[0] * cropFactor, (yMax + yMin) / 2, -220],
-      'x'
-    );
-    this.addSphere(
-      viewport,
-      [xMax - worldDimensions[0] * cropFactor, (yMax + yMin) / 2, -220],
-      'x'
-    );
-    this.addSphere(
-      viewport,
-      [(xMax + xMin) / 2, yMin + worldDimensions[0] * cropFactor, -220],
-      'y'
-    );
-
-    this.addSphere(
-      viewport,
-      [(xMax + xMin) / 2, yMax - worldDimensions[0] * cropFactor, -220],
-      'y'
-    );
-
-    this.addSphere(
-      viewport,
-      [(xMax + xMin) / 2, (yMax + yMin) / 2, -zMax],
-      'z'
-    );
-    this.addSphere(
-      viewport,
-      [(xMax + xMin) / 2, (yMax + yMin) / 2, zMax / 4],
-      'z'
-    );
+    this.addSphere(viewport, sphereXminPoint, 'x');
+    this.addSphere(viewport, sphereXmaxPoint, 'x');
+    this.addSphere(viewport, sphereYminPoint, 'y');
+    this.addSphere(viewport, sphereYmaxPoint, 'y');
+    this.addSphere(viewport, sphereZminPoint, 'z');
+    this.addSphere(viewport, sphereZmaxPoint, 'z');
     const defaultActor = viewport.getDefaultActor();
     if (defaultActor?.actor) {
       // Cast to any to avoid type errors with different actor types
@@ -576,16 +405,38 @@ class VolumeCroppingTool extends AnnotationTool {
     element.addEventListener('mousedown', this._onMouseDownSphere);
     element.addEventListener('mousemove', this._onMouseMoveSphere);
     element.addEventListener('mouseup', this._onMouseUpSphere);
-    //  mapper.modified();
-    //  viewport.getDefaultActor().actor.modified();
+    mapper.modified();
+    viewport.getDefaultActor().actor.modified();
+
+    eventTarget.addEventListener(
+      Events.CROSSHAIR_TOOL_CENTER_CHANGED,
+      (evt) => {
+        console.debug('CROSSHAIR_TOOL_CENTER_CHANGED', evt);
+        viewportsInfo = this._getViewportsInfo();
+        const [viewport3D] = viewportsInfo;
+
+        const renderingEngine = getRenderingEngine(
+          viewport3D.renderingEngineId
+        );
+        const viewport = renderingEngine.getViewport(viewport3D.viewportId);
+
+        const { toolCenter } = evt.detail;
+        viewport.setCamera({
+          focalPoint: toolCenter,
+        });
+      }
+    );
+
+    eventTarget.addEventListener(Events.VOLUMECROPPING_TOOL_CHANGED, (evt) => {
+      console.debug('VOLUMECROPPING_TOOL_CHANGED', evt.data);
+      viewportsInfo = this._getViewportsInfo();
+      const [viewport3D] = viewportsInfo;
+
+      const renderingEngine = getRenderingEngine(viewport3D.renderingEngineId);
+      const viewport = renderingEngine.getViewport(viewport3D.viewportId);
+    });
 
     viewport.render();
-    const toolCenter = csUtils.planar.threePlaneIntersection(
-      firstPlane,
-      secondPlane,
-      thirdPlane
-    );
-    this.setToolCenter(toolCenter);
   };
 
   /**
@@ -635,8 +486,7 @@ class VolumeCroppingTool extends AnnotationTool {
   _onMouseDownSphere = (evt) => {
     const element = evt.currentTarget;
     const viewportsInfo = this._getViewportsInfo();
-    const [firstViewport, secondViewport, thirdViewport, viewport3D] =
-      viewportsInfo;
+    const [viewport3D] = viewportsInfo;
 
     const renderingEngine = getRenderingEngine(viewport3D.renderingEngineId);
     const viewport = renderingEngine.getViewport(viewport3D.viewportId);
@@ -652,7 +502,7 @@ class VolumeCroppingTool extends AnnotationTool {
         // 20 pixels threshold
         this.draggingSphereIndex = i;
         element.style.cursor = 'grabbing';
-        console.debug('grabbing', i);
+        console.debug('grabbing sphere index: ', i);
         return;
       }
     }
@@ -668,8 +518,7 @@ class VolumeCroppingTool extends AnnotationTool {
 
     const element = evt.currentTarget;
     const viewportsInfo = this._getViewportsInfo();
-    const [firstViewport, secondViewport, thirdViewport, viewport3D] =
-      viewportsInfo;
+    const [viewport3D] = viewportsInfo;
     const renderingEngine = getRenderingEngine(viewport3D.renderingEngineId);
     const viewport = renderingEngine.getViewport(viewport3D.viewportId);
 
@@ -713,7 +562,7 @@ class VolumeCroppingTool extends AnnotationTool {
 
       const clippingPlanes = mapper.getClippingPlanes();
       //console.debug('clippingPlanes before setOrigin:', clippingPlanes);
-      //clippingPlanes[this.draggingSphereIndex].setOrigin(newPoint);
+      clippingPlanes[this.draggingSphereIndex].setOrigin(newPoint);
       viewport.setOriginalClippingPlane(this.draggingSphereIndex, newPoint);
       mapper.modified();
       viewport.getDefaultActor().actor.modified();
@@ -728,98 +577,6 @@ class VolumeCroppingTool extends AnnotationTool {
     evt.currentTarget.style.cursor = '';
     //   }
     this.draggingSphereIndex = null;
-  };
-
-  setToolCenter(toolCenter: Types.Point3, suppressEvents = false): void {
-    // prettier-ignore
-    this.toolCenter = toolCenter;
-    const viewportsInfo = this._getViewportsInfo();
-
-    // assuming all viewports are in the same rendering engine
-    triggerAnnotationRenderForViewportIds(
-      viewportsInfo.map(({ viewportId }) => viewportId)
-    );
-    if (!suppressEvents) {
-      triggerEvent(eventTarget, Events.CROSSHAIR_TOOL_CENTER_CHANGED, {
-        toolGroupId: this.toolGroupId,
-        toolCenter: this.toolCenter,
-      });
-    }
-  }
-
-  /**
-   * addNewAnnotation acts as jump for the crosshairs tool. It is called when
-   * the user clicks on the image. It does not store the annotation in the stateManager though.
-   *
-   * @param evt - The mouse event
-   * @param interactionType - The type of interaction (e.g., mouse, touch, etc.)
-   * @returns Crosshairs annotation
-   */
-
-  addNewAnnotation = (
-    evt: EventTypes.InteractionEventType
-  ): VolumeCroppingAnnotation => {
-    const eventDetail = evt.detail;
-    const { element } = eventDetail;
-
-    const { currentPoints } = eventDetail;
-    const jumpWorld = currentPoints.world;
-
-    const enabledElement = getEnabledElement(element);
-    const { viewport } = enabledElement;
-    if (viewport.type === Enums.ViewportType.VOLUME_3D) {
-      const annotations = this._getAnnotations(enabledElement);
-      const filteredAnnotations = this.filterInteractableAnnotationsForElement(
-        viewport.element,
-        annotations
-      );
-
-      const { data } = filteredAnnotations[0];
-      data.handles.activeOperation = OPERATION.DRAG;
-      return filteredAnnotations[0];
-    } else {
-      this._jump(enabledElement, jumpWorld);
-
-      const annotations = this._getAnnotations(enabledElement);
-      const filteredAnnotations = this.filterInteractableAnnotationsForElement(
-        viewport.element,
-        annotations
-      );
-
-      const { data } = filteredAnnotations[0];
-
-      const viewportIdArray = [];
-      // put all the draggable reference lines in the viewportIdArray
-      /*
-    for (let i = 0; i < rotationPoints.length - 1; ++i) {
-      const otherViewport = rotationPoints[i][1];
-      const viewportControllable = this._getReferenceLineControllable(
-        otherViewport.id
-      );
-
-      if (!viewportControllable) {
-        continue;
-      }
-      viewportIdArray.push(otherViewport.id);
-      // rotation handles are two per viewport
-      i++;
-    }
-*/
-      data.activeViewportIds = [...viewportIdArray];
-      // set translation operation
-      data.handles.activeOperation = OPERATION.DRAG;
-
-      evt.preventDefault();
-
-      hideElementCursor(element);
-
-      this._activateModify(element);
-      return filteredAnnotations[0];
-    }
-  };
-
-  cancel = () => {
-    console.log('Not implemented yet');
   };
 
   /**
@@ -846,142 +603,27 @@ class VolumeCroppingTool extends AnnotationTool {
     return false;
   };
 
-  toolSelectedCallback = (
-    evt: EventTypes.InteractionEventType,
-    annotation: Annotation,
-    interactionType: InteractionTypes
-  ): void => {
-    const eventDetail = evt.detail;
-    const { element } = eventDetail;
-    annotation.highlighted = true;
-    this._activateModify(element);
-
-    hideElementCursor(element);
-
-    evt.preventDefault();
-  };
-
   onCameraModified = (evt) => {
-    const eventDetail = evt.detail;
-    const { element } = eventDetail;
+    const { element } = evt.currentTarget
+      ? { element: evt.currentTarget }
+      : evt.detail;
     const enabledElement = getEnabledElement(element);
-    const { renderingEngine } = enabledElement;
-    const viewport = enabledElement.viewport as Types.IVolumeViewport;
-
-    const annotations = this._getAnnotations(enabledElement);
-    const filteredToolAnnotations =
-      this.filterInteractableAnnotationsForElement(element, annotations);
-
-    // viewport that the camera modified is originating from
-    const viewportAnnotation =
-      filteredToolAnnotations[0] as VolumeCroppingAnnotation;
-
-    if (!viewportAnnotation) {
+    const { viewport } = enabledElement;
+    const volumeActor = viewport.getDefaultActor()?.actor;
+    if (!volumeActor) {
+      console.warn('No volume actor found');
       return;
     }
+    const mapper = volumeActor.getMapper();
 
-    // -- Update the camera of other linked viewports containing the same volumeId that
-    //    have the same camera in case of translation
-    // -- Update the crosshair center in world coordinates in annotation.
-    // This is necessary because other tools can modify the position of the slices,
-    // e.g. stackScroll tool at wheel scroll. So we update the coordinates of the center always here.
-    // NOTE: rotation and slab thickness handles are created/updated in renderTool.
-    const currentCamera = viewport.getCamera();
-    const oldCameraPosition = viewportAnnotation.metadata.cameraPosition;
-    const deltaCameraPosition: Types.Point3 = [0, 0, 0];
-    vtkMath.subtract(
-      currentCamera.position,
-      oldCameraPosition,
-      deltaCameraPosition
-    );
+    const clippingPlanes = mapper.getClippingPlanes();
 
-    const oldCameraFocalPoint = viewportAnnotation.metadata.cameraFocalPoint;
-    const deltaCameraFocalPoint: Types.Point3 = [0, 0, 0];
-    vtkMath.subtract(
-      currentCamera.focalPoint,
-      oldCameraFocalPoint,
-      deltaCameraFocalPoint
-    );
-
-    // updated cached "previous" camera position and focal point
-    viewportAnnotation.metadata.cameraPosition = [...currentCamera.position];
-    viewportAnnotation.metadata.cameraFocalPoint = [
-      ...currentCamera.focalPoint,
-    ];
-
-    const viewportControllable = this._getReferenceLineControllable(
-      viewport.id
-    );
-
-    if (
-      !csUtils.isEqual(currentCamera.position, oldCameraPosition, 1e-3) &&
-      viewportControllable
-    ) {
-      // Is camera Modified a TRANSLATION or ROTATION?
-      let isRotation = false;
-
-      // This is guaranteed to be the same diff for both position and focal point
-      // if the camera is modified by pan, zoom, or scroll BUT for rotation of
-      // crosshairs handles it will be different.
-      const cameraModifiedSameForPosAndFocalPoint = csUtils.isEqual(
-        deltaCameraPosition,
-        deltaCameraFocalPoint,
-        1e-3
-      );
-
-      // NOTE: it is a translation if the the focal point and camera position shifts are the same
-      if (!cameraModifiedSameForPosAndFocalPoint) {
-        isRotation = true;
-      }
-
-      const cameraModifiedInPlane =
-        Math.abs(
-          vtkMath.dot(deltaCameraPosition, currentCamera.viewPlaneNormal)
-        ) < 1e-2;
-
-      // TRANSLATION
-      // NOTE1: if the camera modified is a result of a pan or zoom don't update the crosshair center
-      // NOTE2: rotation handles are updates in renderTool
-      if (!isRotation && !cameraModifiedInPlane) {
-        this.toolCenter[0] += deltaCameraPosition[0];
-        this.toolCenter[1] += deltaCameraPosition[1];
-        this.toolCenter[2] += deltaCameraPosition[2];
-
-        triggerEvent(eventTarget, Events.CROSSHAIR_TOOL_CENTER_CHANGED, {
-          toolGroupId: this.toolGroupId,
-          toolCenter: this.toolCenter,
-        });
-      }
-    }
-
-    // AutoPan modification
-    if (this.configuration.autoPan?.enabled) {
-      const toolGroup = getToolGroupForViewport(
-        viewport.id,
-        renderingEngine.id
-      );
-
-      const otherViewportIds = toolGroup
-        .getViewportIds()
-        .filter((id) => id !== viewport.id);
-
-      otherViewportIds.forEach((viewportId) => {
-        this._autoPanViewportIfNecessary(viewportId, renderingEngine);
-      });
-    }
-
-    const requireSameOrientation = false;
-    const viewportIdsToRender = getViewportIdsWithToolToRender(
-      element,
-      this.getToolName(),
-      requireSameOrientation
-    );
-
-    triggerAnnotationRenderForViewportIds(viewportIdsToRender);
+    // console.debug('on camera modified', viewport.getActors(), clippingPlanes);
+    enabledElement.viewport.render();
   };
 
   onResetCamera = (evt) => {
-    this.resetCrosshairs();
+    console.debug('on reset camera');
   };
 
   mouseMoveCallback = (
@@ -1015,23 +657,8 @@ class VolumeCroppingTool extends AnnotationTool {
 
       // This init are necessary, because when we move the mouse they are not cleaned by _endCallback
       data.activeViewportIds = [];
-
-      /*
-      data.handles.activeOperation = null;
-
-      const handleNearImagePoint = this.getHandleNearImagePoint(
-        element,
-        annotation,
-        canvasCoords,
-        6
-      );
-*/
       let near = false;
-      //     if (handleNearImagePoint) {
-      //      near = true;
-      //    } else {
       near = this._pointNearTool(element, annotation, canvasCoords, 6);
-      //   }
 
       const nearToolAndNotMarkedActive = near && !highlighted;
       const notNearToolAndMarkedActive = !near && highlighted;
@@ -1044,370 +671,24 @@ class VolumeCroppingTool extends AnnotationTool {
     return imageNeedsUpdate;
   };
 
-  filterInteractableAnnotationsForElement = (element, annotations) => {
-    if (!annotations || !annotations.length) {
-      return [];
-    }
-
-    const enabledElement = getEnabledElement(element);
-    const { viewportId } = enabledElement;
-    // console.log(annotations);
-    const viewportUIDSpecificCrosshairs = annotations.filter(
-      (annotation) => annotation.data.viewportId === viewportId
-    );
-
-    return viewportUIDSpecificCrosshairs;
-  };
-
-  /**
-   * renders the volume cropping lines and handles in the requestAnimationFrame callback
-   *
-   * @param enabledElement - The Cornerstone's enabledElement.
-   * @param svgDrawingHelper - The svgDrawingHelper providing the context for drawing.
-   */
-  renderAnnotation = (
-    enabledElement: Types.IEnabledElement,
-    svgDrawingHelper: SVGDrawingHelper
-  ): boolean => {
-    let renderStatus = false;
-    const { viewport, renderingEngine } = enabledElement;
-    const { element } = viewport;
-    const annotations = this._getAnnotations(enabledElement);
-    const camera = viewport.getCamera();
-    const filteredToolAnnotations =
-      this.filterInteractableAnnotationsForElement(element, annotations);
-
-    // viewport Annotation
-    const viewportAnnotation = filteredToolAnnotations[0];
-    if (!annotations?.length || !viewportAnnotation?.data) {
-      // No annotations yet, and didn't just create it as we likely don't have a FrameOfReference/any data loaded yet.
-      return renderStatus;
-    }
-
-    const annotationUID = viewportAnnotation.annotationUID;
-
-    // Get cameras/canvases for each of these.
-    // -- Get two world positions for this canvas in this line (e.g. the diagonal)
-    // -- Convert these world positions to this canvas.
-    // -- Extend/confine this line to fit in this canvas.
-    // -- Render this line.
-    const { clientWidth, clientHeight } = viewport.canvas;
-    const canvasDiagonalLength = Math.sqrt(
-      clientWidth * clientWidth + clientHeight * clientHeight
-    );
-    const canvasMinDimensionLength = Math.min(clientWidth, clientHeight);
-
-    const data = viewportAnnotation.data;
-    const crosshairCenterCanvas = viewport.worldToCanvas(this.toolCenter);
-
-    const otherViewportAnnotations =
-      this._filterAnnotationsByUniqueViewportOrientations(
-        enabledElement,
-        annotations
-      );
-
-    const referenceLines = [];
-
-    // get canvas information for points and lines (canvas box, canvas horizontal distances)
-    const canvasBox = [0, 0, clientWidth, clientHeight];
-
-    otherViewportAnnotations.forEach((annotation) => {
-      const { data } = annotation;
-
-      data.handles.toolCenter = this.toolCenter;
-
-      const otherViewport = renderingEngine.getViewport(
-        data.viewportId
-      ) as Types.IVolumeViewport;
-
-      const otherCamera = otherViewport.getCamera();
-
-      const otherViewportControllable = this._getReferenceLineControllable(
-        otherViewport.id
-      );
-
-      // get coordinates for the reference line
-      const { clientWidth, clientHeight } = otherViewport.canvas;
-      const otherCanvasDiagonalLength = Math.sqrt(
-        clientWidth * clientWidth + clientHeight * clientHeight
-      );
-      const otherCanvasCenter: Types.Point2 = [
-        clientWidth * 0.5,
-        clientHeight * 0.5,
-      ];
-      const otherViewportCenterWorld =
-        otherViewport.canvasToWorld(otherCanvasCenter);
-
-      const direction: Types.Point3 = [0, 0, 0];
-      vtkMath.cross(
-        camera.viewPlaneNormal,
-        otherCamera.viewPlaneNormal,
-        direction
-      );
-      vtkMath.normalize(direction);
-      vtkMath.multiplyScalar(
-        <Types.Point3>direction,
-        otherCanvasDiagonalLength
-      );
-
-      const pointWorld0: Types.Point3 = [0, 0, 0];
-      vtkMath.add(otherViewportCenterWorld, direction, pointWorld0);
-
-      const pointWorld1: Types.Point3 = [0, 0, 0];
-      vtkMath.subtract(otherViewportCenterWorld, direction, pointWorld1);
-
-      const pointCanvas0 = viewport.worldToCanvas(pointWorld0);
-
-      const otherViewportCenterCanvas = viewport.worldToCanvas(
-        otherViewportCenterWorld
-      );
-
-      const canvasUnitVectorFromCenter = vec2.create();
-      vec2.subtract(
-        canvasUnitVectorFromCenter,
-        pointCanvas0,
-        otherViewportCenterCanvas
-      );
-      vec2.normalize(canvasUnitVectorFromCenter, canvasUnitVectorFromCenter);
-
-      // Graphic:
-      // Mid -> SlabThickness handle
-      // Short -> Rotation handle
-      //                           Long
-      //                            |
-      //                            |
-      //                            |
-      //                           Mid
-      //                            |
-      //                            |
-      //                            |
-      //                          Short
-      //                            |
-      //                            |
-      //                            |
-      // Long --- Mid--- Short--- Center --- Short --- Mid --- Long
-      //                            |
-      //                            |
-      //                            |
-      //                          Short
-      //                            |
-      //                            |
-      //                            |
-      //                           Mid
-      //                            |
-      //                            |
-      //                            |
-      //                           Long
-      const canvasVectorFromCenterLong = vec2.create();
-
-      vec2.scale(
-        canvasVectorFromCenterLong,
-        canvasUnitVectorFromCenter,
-        canvasDiagonalLength * 100
-      );
-      const canvasVectorFromCenterMid = vec2.create();
-      vec2.scale(
-        canvasVectorFromCenterMid,
-        canvasUnitVectorFromCenter,
-        // to maximize the visibility of the controls, they need to be
-        // placed at most at half the length of the shortest side of the canvas.
-        // Chosen 0.4 to have some margin to the edge.
-        canvasMinDimensionLength * 0.4
-      );
-      const canvasVectorFromCenterShort = vec2.create();
-      vec2.scale(
-        canvasVectorFromCenterShort,
-        canvasUnitVectorFromCenter,
-        // Chosen 0.2 because is half of 0.4.
-        canvasMinDimensionLength * 0.2
-      );
-      const canvasVectorFromCenterStart = vec2.create();
-      const centerGap = this.configuration.referenceLinesCenterGapRadius;
-      vec2.scale(
-        canvasVectorFromCenterStart,
-        canvasUnitVectorFromCenter,
-        // Don't put a gap if the the third view is missing
-        otherViewportAnnotations.length === 2 ? centerGap : 0
-      );
-
-      // Computing Reference start and end (4 lines per viewport in case of 3 view MPR)
-      const refLinePointOne = vec2.create();
-      const refLinePointTwo = vec2.create();
-      const refLinePointThree = vec2.create();
-      const refLinePointFour = vec2.create();
-
-      let refLinesCenter = vec2.clone(crosshairCenterCanvas);
-      if (!otherViewportControllable) {
-        refLinesCenter = vec2.clone(otherViewportCenterCanvas);
-      }
-
-      vec2.add(refLinePointOne, refLinesCenter, canvasVectorFromCenterStart);
-      vec2.add(refLinePointTwo, refLinesCenter, canvasVectorFromCenterLong);
-      vec2.subtract(
-        refLinePointThree,
-        refLinesCenter,
-        canvasVectorFromCenterStart
-      );
-      vec2.subtract(
-        refLinePointFour,
-        refLinesCenter,
-        canvasVectorFromCenterLong
-      );
-
-      // Clipping lines to be only included in a box (canvas), we don't want
-      // the lines goes beyond canvas
-      liangBarksyClip(refLinePointOne, refLinePointTwo, canvasBox);
-      liangBarksyClip(refLinePointThree, refLinePointFour, canvasBox);
-      referenceLines.push([
-        otherViewport,
-        refLinePointOne,
-        refLinePointTwo,
-        refLinePointThree,
-        refLinePointFour,
-      ]);
-    });
-
-    if (viewport.type !== Enums.ViewportType.VOLUME_3D) {
-      data.referenceLines = referenceLines;
-      const viewportColor = this._getReferenceLineColor(viewport.id);
-      const color =
-        viewportColor !== undefined ? viewportColor : 'rgb(200, 200, 200)';
-
-      referenceLines.forEach((line, lineIndex) => {
-        // get color for the reference line
-        const otherViewport = line[0];
-        const viewportColor = this._getReferenceLineColor(otherViewport.id);
-        const viewportControllable = this._getReferenceLineControllable(
-          otherViewport.id
-        );
-        const selectedViewportId = data.activeViewportIds.find(
-          (id) => id === otherViewport.id
-        );
-
-        const color =
-          viewportColor !== undefined ? viewportColor : 'rgb(200, 200, 200)';
-
-        let lineWidth = 1;
-
-        const lineActive =
-          data.handles.activeOperation !== null &&
-          data.handles.activeOperation === OPERATION.DRAG &&
-          selectedViewportId;
-
-        if (lineActive) {
-          lineWidth = 2.5;
-        }
-
-        let lineUID = `${lineIndex}`;
-        if (viewportControllable) {
-          lineUID = `${lineIndex}One`;
-          console.debug(lineUID);
-          drawLineSvg(
-            svgDrawingHelper,
-            annotationUID,
-            lineUID,
-            line[1],
-            line[2],
-            {
-              color,
-              lineWidth,
-            }
-          );
-
-          lineUID = `${lineIndex}Two`;
-          drawLineSvg(
-            svgDrawingHelper,
-            annotationUID,
-            lineUID,
-            line[3],
-            line[4],
-            {
-              color,
-              lineWidth,
-            }
-          );
-        } else {
-          drawLineSvg(
-            svgDrawingHelper,
-            annotationUID,
-            lineUID,
-            line[2],
-            line[4],
-            {
-              color,
-              lineWidth,
-            }
-          );
-        }
-
-        if (viewportControllable) {
-          color =
-            viewportColor !== undefined ? viewportColor : 'rgb(200, 200, 200)';
-          let handleRadius =
-            this.configuration.handleRadius *
-            (this.configuration.enableHDPIHandles
-              ? window.devicePixelRatio
-              : 1);
-          let opacity = 1;
-          if (this.configuration.mobile?.enabled) {
-            handleRadius = this.configuration.mobile.handleRadius;
-            opacity = this.configuration.mobile.opacity;
-          }
-          if (lineActive) {
-            const handleUID = `${lineIndex}`;
-          }
-        }
-      });
-    }
-    renderStatus = true;
-
-    if (this.configuration.viewportIndicators) {
-      const { viewportIndicatorsConfig } = this.configuration;
-
-      const xOffset = viewportIndicatorsConfig?.xOffset || 0.95;
-      const yOffset = viewportIndicatorsConfig?.yOffset || 0.05;
-      const referenceColorCoordinates = [
-        clientWidth * xOffset,
-        clientHeight * yOffset,
-      ];
-
-      const circleRadius =
-        viewportIndicatorsConfig?.circleRadius || canvasDiagonalLength * 0.01;
-
-      const circleUID = '0';
-      drawCircleSvg(
-        svgDrawingHelper,
-        annotationUID,
-        circleUID,
-        referenceColorCoordinates as Types.Point2,
-        circleRadius,
-        { color, fill: color }
-      );
-    }
-
-    return renderStatus;
-  };
-
-  _getAnnotations = (enabledElement: Types.IEnabledElement) => {
-    const { viewport } = enabledElement;
-    const annotations =
-      getAnnotations(this.getToolName(), viewport.element) || [];
-    const viewportIds = this._getViewportsInfo().map(
-      ({ viewportId }) => viewportId
-    );
-
-    // filter the annotations to only keep that are for this toolGroup
-    const toolGroupAnnotations = annotations.filter((annotation) => {
-      const { data } = annotation;
-      return viewportIds.includes(data.viewportId);
-    });
-
-    return toolGroupAnnotations;
-  };
-
   _onNewVolume = () => {
     const viewportsInfo = this._getViewportsInfo();
-    this._computeToolCenter(viewportsInfo);
+    this._initialize3DViewports(viewportsInfo);
+
+    const [viewport3D] = viewportsInfo;
+    const renderingEngine = getRenderingEngine(viewport3D.renderingEngineId);
+    const viewport = renderingEngine.getViewport(viewport3D.viewportId);
+
+    const camera = viewport.getCamera();
+    viewport.setCamera({
+      focalPoint: camera.focalPoint,
+      position: [
+        camera.position[0],
+        camera.position[1],
+        camera.position[2] + 1,
+      ],
+    });
+    viewport.render();
   };
 
   _unsubscribeToViewportNewVolumeSet(viewportsInfo) {
@@ -1439,443 +720,6 @@ class VolumeCroppingTool extends AnnotationTool {
       );
     });
   }
-
-  _autoPanViewportIfNecessary(
-    viewportId: string,
-    renderingEngine: Types.IRenderingEngine
-  ): void {
-    // 1. Check if the toolCenter is outside the viewport
-    // 2. If it is outside, pan the viewport to fit in the toolCenter
-
-    const viewport = renderingEngine.getViewport(viewportId);
-    const { clientWidth, clientHeight } = viewport.canvas;
-
-    const toolCenterCanvas = viewport.worldToCanvas(this.toolCenter);
-
-    // pan the viewport to fit the toolCenter in the direction
-    // that is out of bounds
-    const pan = this.configuration.autoPan.panSize;
-
-    const visiblePointCanvas = <Types.Point2>[
-      toolCenterCanvas[0],
-      toolCenterCanvas[1],
-    ];
-
-    if (toolCenterCanvas[0] < 0) {
-      visiblePointCanvas[0] = pan;
-    } else if (toolCenterCanvas[0] > clientWidth) {
-      visiblePointCanvas[0] = clientWidth - pan;
-    }
-
-    if (toolCenterCanvas[1] < 0) {
-      visiblePointCanvas[1] = pan;
-    } else if (toolCenterCanvas[1] > clientHeight) {
-      visiblePointCanvas[1] = clientHeight - pan;
-    }
-
-    if (
-      visiblePointCanvas[0] === toolCenterCanvas[0] &&
-      visiblePointCanvas[1] === toolCenterCanvas[1]
-    ) {
-      return;
-    }
-
-    const visiblePointWorld = viewport.canvasToWorld(visiblePointCanvas);
-
-    const deltaPointsWorld = [
-      visiblePointWorld[0] - this.toolCenter[0],
-      visiblePointWorld[1] - this.toolCenter[1],
-      visiblePointWorld[2] - this.toolCenter[2],
-    ];
-
-    const camera = viewport.getCamera();
-    const { focalPoint, position } = camera;
-
-    const updatedPosition = <Types.Point3>[
-      position[0] - deltaPointsWorld[0],
-      position[1] - deltaPointsWorld[1],
-      position[2] - deltaPointsWorld[2],
-    ];
-
-    const updatedFocalPoint = <Types.Point3>[
-      focalPoint[0] - deltaPointsWorld[0],
-      focalPoint[1] - deltaPointsWorld[1],
-      focalPoint[2] - deltaPointsWorld[2],
-    ];
-
-    viewport.setCamera({
-      focalPoint: updatedFocalPoint,
-      position: updatedPosition,
-    });
-
-    viewport.render();
-  }
-
-  _areViewportIdArraysEqual = (viewportIdArrayOne, viewportIdArrayTwo) => {
-    if (viewportIdArrayOne.length !== viewportIdArrayTwo.length) {
-      return false;
-    }
-
-    viewportIdArrayOne.forEach((id) => {
-      let itemFound = false;
-      for (let i = 0; i < viewportIdArrayTwo.length; ++i) {
-        if (id === viewportIdArrayTwo[i]) {
-          itemFound = true;
-          break;
-        }
-      }
-      if (itemFound === false) {
-        return false;
-      }
-    });
-
-    return true;
-  };
-
-  // It filters the viewports with crosshairs and only return viewports
-  // that have different camera.
-  _getAnnotationsForViewportsWithDifferentCameras = (
-    enabledElement,
-    annotations
-  ) => {
-    const { viewportId, renderingEngine, viewport } = enabledElement;
-
-    const otherViewportAnnotations = annotations.filter(
-      (annotation) => annotation.data.viewportId !== viewportId
-    );
-
-    if (!otherViewportAnnotations || !otherViewportAnnotations.length) {
-      return [];
-    }
-
-    const camera = viewport.getCamera();
-    const { viewPlaneNormal, position } = camera;
-
-    const viewportsWithDifferentCameras = otherViewportAnnotations.filter(
-      (annotation) => {
-        const { viewportId } = annotation.data;
-        const targetViewport = renderingEngine.getViewport(viewportId);
-        const cameraOfTarget = targetViewport.getCamera();
-
-        return !(
-          csUtils.isEqual(
-            cameraOfTarget.viewPlaneNormal,
-            viewPlaneNormal,
-            1e-2
-          ) && csUtils.isEqual(cameraOfTarget.position, position, 1)
-        );
-      }
-    );
-
-    return viewportsWithDifferentCameras;
-  };
-
-  _filterViewportWithSameOrientation = (
-    enabledElement,
-    referenceAnnotation,
-    annotations
-  ) => {
-    const { renderingEngine } = enabledElement;
-    const { data } = referenceAnnotation;
-    const viewport = renderingEngine.getViewport(data.viewportId);
-
-    const linkedViewportAnnotations = annotations.filter((annotation) => {
-      const { data } = annotation;
-      const otherViewport = renderingEngine.getViewport(data.viewportId);
-      const otherViewportControllable = this._getReferenceLineControllable(
-        otherViewport.id
-      );
-
-      return otherViewportControllable === true;
-    });
-
-    if (!linkedViewportAnnotations || !linkedViewportAnnotations.length) {
-      return [];
-    }
-
-    const camera = viewport.getCamera();
-    const viewPlaneNormal = camera.viewPlaneNormal;
-    vtkMath.normalize(viewPlaneNormal);
-
-    const otherViewportsAnnotationsWithSameCameraDirection =
-      linkedViewportAnnotations.filter((annotation) => {
-        const { viewportId } = annotation.data;
-        const otherViewport = renderingEngine.getViewport(viewportId);
-        const otherCamera = otherViewport.getCamera();
-        const otherViewPlaneNormal = otherCamera.viewPlaneNormal;
-        vtkMath.normalize(otherViewPlaneNormal);
-
-        return (
-          csUtils.isEqual(viewPlaneNormal, otherViewPlaneNormal, 1e-2) &&
-          csUtils.isEqual(camera.viewUp, otherCamera.viewUp, 1e-2)
-        );
-      });
-
-    return otherViewportsAnnotationsWithSameCameraDirection;
-  };
-
-  _filterAnnotationsByUniqueViewportOrientations = (
-    enabledElement,
-    annotations
-  ) => {
-    const { renderingEngine, viewport } = enabledElement;
-    const camera = viewport.getCamera();
-    const viewPlaneNormal = camera.viewPlaneNormal;
-    vtkMath.normalize(viewPlaneNormal);
-
-    const otherLinkedViewportAnnotationsFromSameScene = annotations.filter(
-      (annotation) => {
-        const { data } = annotation;
-        const otherViewport = renderingEngine.getViewport(data.viewportId);
-        const otherViewportControllable = this._getReferenceLineControllable(
-          otherViewport.id
-        );
-        // Filter out 3D viewports
-        if (otherViewport.type === Enums.ViewportType.VOLUME_3D) {
-          return false;
-        }
-
-        return (
-          viewport !== otherViewport &&
-          // scene === otherScene &&
-          otherViewportControllable === true
-        );
-      }
-    );
-
-    const otherViewportsAnnotationsWithUniqueCameras = [];
-    // Iterate first on other viewport from the same scene linked
-    for (
-      let i = 0;
-      i < otherLinkedViewportAnnotationsFromSameScene.length;
-      ++i
-    ) {
-      const annotation = otherLinkedViewportAnnotationsFromSameScene[i];
-      const { viewportId } = annotation.data;
-      const otherViewport = renderingEngine.getViewport(viewportId);
-      const otherCamera = otherViewport.getCamera();
-      const otherViewPlaneNormal = otherCamera.viewPlaneNormal;
-      vtkMath.normalize(otherViewPlaneNormal);
-
-      if (
-        csUtils.isEqual(viewPlaneNormal, otherViewPlaneNormal, 1e-2) ||
-        csUtils.isOpposite(viewPlaneNormal, otherViewPlaneNormal, 1e-2)
-      ) {
-        continue;
-      }
-
-      let cameraFound = false;
-      for (
-        let jj = 0;
-        jj < otherViewportsAnnotationsWithUniqueCameras.length;
-        ++jj
-      ) {
-        const annotation = otherViewportsAnnotationsWithUniqueCameras[jj];
-        const { viewportId } = annotation.data;
-        const stockedViewport = renderingEngine.getViewport(viewportId);
-        const cameraOfStocked = stockedViewport.getCamera();
-
-        if (
-          csUtils.isEqual(
-            cameraOfStocked.viewPlaneNormal,
-            otherCamera.viewPlaneNormal,
-            1e-2
-          ) &&
-          csUtils.isEqual(cameraOfStocked.position, otherCamera.position, 1)
-        ) {
-          cameraFound = true;
-        }
-      }
-
-      if (!cameraFound) {
-        otherViewportsAnnotationsWithUniqueCameras.push(annotation);
-      }
-    }
-
-    const otherNonLinkedViewportAnnotationsFromSameScene = annotations.filter(
-      (annotation) => {
-        const { data } = annotation;
-        const otherViewport = renderingEngine.getViewport(data.viewportId);
-        const otherViewportControllable = this._getReferenceLineControllable(
-          otherViewport.id
-        );
-
-        return (
-          viewport !== otherViewport &&
-          // scene === otherScene &&
-          otherViewportControllable !== true
-        );
-      }
-    );
-
-    // Iterate second on other viewport from the same scene non linked
-    for (
-      let i = 0;
-      i < otherNonLinkedViewportAnnotationsFromSameScene.length;
-      ++i
-    ) {
-      const annotation = otherNonLinkedViewportAnnotationsFromSameScene[i];
-      const { viewportId } = annotation.data;
-      const otherViewport = renderingEngine.getViewport(viewportId);
-
-      const otherCamera = otherViewport.getCamera();
-      const otherViewPlaneNormal = otherCamera.viewPlaneNormal;
-      vtkMath.normalize(otherViewPlaneNormal);
-
-      if (
-        csUtils.isEqual(viewPlaneNormal, otherViewPlaneNormal, 1e-2) ||
-        csUtils.isOpposite(viewPlaneNormal, otherViewPlaneNormal, 1e-2)
-      ) {
-        continue;
-      }
-
-      let cameraFound = false;
-      for (
-        let jj = 0;
-        jj < otherViewportsAnnotationsWithUniqueCameras.length;
-        ++jj
-      ) {
-        const annotation = otherViewportsAnnotationsWithUniqueCameras[jj];
-        const { viewportId } = annotation.data;
-        const stockedViewport = renderingEngine.getViewport(viewportId);
-        const cameraOfStocked = stockedViewport.getCamera();
-
-        if (
-          csUtils.isEqual(
-            cameraOfStocked.viewPlaneNormal,
-            otherCamera.viewPlaneNormal,
-            1e-2
-          ) &&
-          csUtils.isEqual(cameraOfStocked.position, otherCamera.position, 1)
-        ) {
-          cameraFound = true;
-        }
-      }
-
-      if (!cameraFound) {
-        otherViewportsAnnotationsWithUniqueCameras.push(annotation);
-      }
-    }
-
-    // Iterate on all the viewport
-    const otherViewportAnnotations =
-      this._getAnnotationsForViewportsWithDifferentCameras(
-        enabledElement,
-        annotations
-      );
-
-    for (let i = 0; i < otherViewportAnnotations.length; ++i) {
-      const annotation = otherViewportAnnotations[i];
-      if (
-        otherViewportsAnnotationsWithUniqueCameras.some(
-          (element) => element === annotation
-        )
-      ) {
-        continue;
-      }
-
-      const { viewportId } = annotation.data;
-      const otherViewport = renderingEngine.getViewport(viewportId);
-      const otherCamera = otherViewport.getCamera();
-      const otherViewPlaneNormal = otherCamera.viewPlaneNormal;
-      vtkMath.normalize(otherViewPlaneNormal);
-
-      if (
-        csUtils.isEqual(viewPlaneNormal, otherViewPlaneNormal, 1e-2) ||
-        csUtils.isOpposite(viewPlaneNormal, otherViewPlaneNormal, 1e-2)
-      ) {
-        continue;
-      }
-
-      let cameraFound = false;
-      for (
-        let jj = 0;
-        jj < otherViewportsAnnotationsWithUniqueCameras.length;
-        ++jj
-      ) {
-        const annotation = otherViewportsAnnotationsWithUniqueCameras[jj];
-        const { viewportId } = annotation.data;
-        const stockedViewport = renderingEngine.getViewport(viewportId);
-        const cameraOfStocked = stockedViewport.getCamera();
-
-        if (
-          csUtils.isEqual(
-            cameraOfStocked.viewPlaneNormal,
-            otherCamera.viewPlaneNormal,
-            1e-2
-          ) &&
-          csUtils.isEqual(cameraOfStocked.position, otherCamera.position, 1)
-        ) {
-          cameraFound = true;
-        }
-      }
-
-      if (!cameraFound) {
-        otherViewportsAnnotationsWithUniqueCameras.push(annotation);
-      }
-    }
-
-    return otherViewportsAnnotationsWithUniqueCameras;
-  };
-
-  _checkIfViewportsRenderingSameScene = (viewport, otherViewport) => {
-    const volumeIds = viewport.getAllVolumeIds();
-    const otherVolumeIds = otherViewport.getAllVolumeIds();
-
-    return (
-      volumeIds.length === otherVolumeIds.length &&
-      volumeIds.every((id) => otherVolumeIds.includes(id))
-    );
-  };
-
-  _jump = (enabledElement, jumpWorld) => {
-    state.isInteractingWithTool = true;
-    const { viewport, renderingEngine } = enabledElement;
-
-    const annotations = this._getAnnotations(enabledElement);
-
-    const delta: Types.Point3 = [0, 0, 0];
-    vtkMath.subtract(jumpWorld, this.toolCenter, delta);
-
-    // TRANSLATION
-    // get the annotation of the other viewport which are parallel to the delta shift and are of the same scene
-    const otherViewportAnnotations =
-      this._getAnnotationsForViewportsWithDifferentCameras(
-        enabledElement,
-        annotations
-      );
-
-    const viewportsAnnotationsToUpdate = otherViewportAnnotations.filter(
-      (annotation) => {
-        const { data } = annotation;
-        const otherViewport = renderingEngine.getViewport(data.viewportId);
-
-        const sameScene = this._checkIfViewportsRenderingSameScene(
-          viewport,
-          otherViewport
-        );
-
-        return (
-          this._getReferenceLineControllable(otherViewport.id) && sameScene
-        );
-      }
-    );
-
-    if (viewportsAnnotationsToUpdate.length === 0) {
-      state.isInteractingWithTool = false;
-      return false;
-    }
-
-    this._applyDeltaShiftToSelectedViewportCameras(
-      renderingEngine,
-      viewportsAnnotationsToUpdate,
-      delta
-    );
-
-    state.isInteractingWithTool = false;
-
-    return true;
-  };
 
   _activateModify = (element) => {
     // mobile sometimes has lingering interaction even when touchEnd triggers
@@ -1970,6 +814,7 @@ class VolumeCroppingTool extends AnnotationTool {
       triggerAnnotationRenderForViewportIds(
         viewportsInfo.map(({ viewportId }) => viewportId)
       );
+      viewport.render;
     }
 
     // TRANSLATION
@@ -2003,58 +848,6 @@ class VolumeCroppingTool extends AnnotationTool {
       delta
     );
   };
-
-  _applyDeltaShiftToSelectedViewportCameras(
-    renderingEngine,
-    viewportsAnnotationsToUpdate,
-    delta
-  ) {
-    // update camera for the other viewports.
-    // NOTE1: The lines then are rendered by the onCameraModified
-    // NOTE2: crosshair center are automatically updated in the onCameraModified event
-    viewportsAnnotationsToUpdate.forEach((annotation) => {
-      this._applyDeltaShiftToViewportCamera(renderingEngine, annotation, delta);
-    });
-  }
-
-  _applyDeltaShiftToViewportCamera(
-    renderingEngine: Types.IRenderingEngine,
-    annotation,
-    delta
-  ) {
-    // update camera for the other viewports.
-    // NOTE1: The lines then are rendered by the onCameraModified
-    // NOTE2: crosshair center are automatically updated in the onCameraModified event
-    const { data } = annotation;
-
-    const viewport = renderingEngine.getViewport(data.viewportId);
-    const camera = viewport.getCamera();
-    const normal = camera.viewPlaneNormal;
-
-    // Project delta over camera normal
-    // (we don't need to pan, we need only to scroll the camera as in the wheel stack scroll tool)
-    const dotProd = vtkMath.dot(delta, normal);
-    const projectedDelta: Types.Point3 = [...normal];
-    vtkMath.multiplyScalar(projectedDelta, dotProd);
-
-    if (
-      Math.abs(projectedDelta[0]) > 1e-3 ||
-      Math.abs(projectedDelta[1]) > 1e-3 ||
-      Math.abs(projectedDelta[2]) > 1e-3
-    ) {
-      const newFocalPoint: Types.Point3 = [0, 0, 0];
-      const newPosition: Types.Point3 = [0, 0, 0];
-
-      vtkMath.add(camera.focalPoint, projectedDelta, newFocalPoint);
-      vtkMath.add(camera.position, projectedDelta, newPosition);
-
-      viewport.setCamera({
-        focalPoint: newFocalPoint,
-        position: newPosition,
-      });
-      viewport.render();
-    }
-  }
 
   _pointNearTool(element, annotation, canvasCoords, proximity) {
     const { data } = annotation;
